add_sleeve_to_images function

+++ dbUtils.py
@@ -1083,13 +1083,16 @@ def add_sleeve_length_to_relevant_items_in_images():
     sent = 0
     deleted = 0
     for doc in db.images.find():
+        if 'sleeve_length' in set([item['fp'].keys() for person in doc for item in person]):
+            continue
         image = Utils.get_cv2_img_array(doc['image_urls'][0])
         if image is None:
             db.images.delete_one({'_id': doc['_id']})
             deleted += 1
             print("{0} images deleted..".format(deleted))
         add_feature.enqueue(parallel_sleeve_and_replace, args=(doc['_id'], image), timeout=2000)
-    sent += 1
+        sent += 1
+        print('Sent {0} docs by now..'.format(sent))
 
 
 def parallel_sleeve_and_replace(image_obj_id, image):
